Cancellation
- We enhanced the **MessagesQueue** API and added the
`addCancelledGroup(String groupId)` method. The cancelled groups are
stored in a internal **Set** and when the `dequeue(String filter)`
method is invoked, we check next matching message against this blacklist
and skip and remove it from the queue if is the case.
- We noticed that we need to remove from the middle of the
**MessagesQueue** for both filtering and canceling operations, so we
choose to replace the internal **List** implementation from
**ArrayList** to **LinkedList** to improve the performance of those
operations.